refactor(session): extract client ChannelStore into its own module
Move the client channel store out of CredentialState.ts into a dedicated
ChannelStore.ts, mirroring session/server/ChannelStore.ts. CredentialState
keeps credential planning; the store owns persistence and serialization.

- new src/tempo/session/client/ChannelStore.ts holds ChannelStore,
  ChannelSink, channelKey/entryKey, createChannelStore, StoredChannel,
  serializeEntry/deserializeEntry, JsonChannelKv, createJsonChannelStore
- drop the store re-export ladder through Session.ts; client barrels import
  straight from ChannelStore.js
- SessionManager and tests import store symbols from the new module

Amp-Thread-ID: https://ampcode.com/threads/T-019ed67a-26f1-725b-8dd8-ad419a893f75
