+++ src/tempo/session/client/SessionManager.test.ts
@@ -6,7 +6,7 @@ import * as Challenge from '../../../Challenge.js'
 import * as Constants from '../../../Constants.js'
 import * as Credential from '../../../Credential.js'
 import type { ChannelEntry } from '../client/ChannelOps.js'
-import { createJsonChannelStore, entryKey, type ChannelStore } from '../client/CredentialState.js'
+import { createJsonChannelStore, entryKey, type ChannelStore } from '../client/ChannelStore.js'
 import * as Channel from '../precompile/Channel.js'
 import { escrowAbi } from '../precompile/escrow.abi.js'
 import { tip20ChannelEscrow } from '../precompile/Protocol.js'
